Many things
Made ducky class that works for a continuous servo, I have tried to move all of the drivebase stuff from the teleop to a class but I'm not sure if it will work. Teleop stuff can be commented out for testing.  I am not sure if the SetRunMode function works like I want it to.